+++ lib/widgets/timeline/tl_key_parts/tl_key.dart
@@ -9,7 +9,7 @@ import 'key_widget.dart';
 
 class TLKey extends StatefulWidget {
   final int frameNumber;
-  final double fps;
+  final int fps;
   final bool isWholeSecond;
 
   const TLKey({
